code beauty

+++ app/Http/Controllers/CourseCreateController.php
@@ -35,26 +35,19 @@ class CourseCreateController extends Controller
     {
         $course_id = $request->session()->get('course_id');
 
-        if (empty($course_id)) {
-            return redirect('/courses/create');
-        }
+        if (empty($course_id)) return redirect('/courses/create');
 
         return view('courses.create.complete', ['course_id' => $course_id]);
     }
 
     public function showTeacherProfileForm(Request $request)
     {
-        // dd($request);
-        // $this->validate($request, [
-        //      'contractCheck' => 'accepted',
-        //  ]);
-
         $signed_contract = $request->session()->get('signed_contract');
-        if (empty($signed_contract)) {
-            return redirect('/courses/create/step/contract');
-        }
+
+        if (empty($signed_contract)) return redirect('/courses/create/step/contract');
 
         $teacher_profile = $request->session()->get('teacher_profile');
+
         if (empty($teacher_profile)) {
             $user = User::find(auth()->user()->id);
             $profile = $user->profile;
@@ -71,9 +64,7 @@ class CourseCreateController extends Controller
             ];
         }
 
-        if(!array_key_exists('avatar', $teacher_profile)) {
-            $teacher_profile['avatar'] = auth()->user()->avatar;
-        }
+        if (!array_key_exists('avatar', $teacher_profile)) $teacher_profile['avatar'] = auth()->user()->avatar;
 
         $request->session()->forget('teacher_profile');
 
@@ -141,7 +132,6 @@ class CourseCreateController extends Controller
             DB::rollback();
             throw $e;
         }
-
         $request->session()->put('save_teacher', true);
         $request->session()->forget('teacher_profile');
 
@@ -197,14 +187,11 @@ class CourseCreateController extends Controller
                 $profile->about = $request['about'];
                 $profile->save();
             }
-
             DB::commit();
-
         } catch (\Exception $e) {
             DB::rollback();
             throw $e;
         }
-
         $request->session()->forget('course');
         $request->session()->forget('save_teacher');
         $request->session()->forget('signed_contract');
@@ -216,9 +203,7 @@ class CourseCreateController extends Controller
     public function showCourseForm(Request $request)
     {
         $save_teacher = $request->session()->get('save_teacher');
-        if (empty($save_teacher)) {
-            return redirect('/courses/create');
-        }
+        if (empty($save_teacher)) return redirect('/courses/create');
 
         $course = $request->session()->get('course');
         $categories = CourseCategory::all();
@@ -233,7 +218,7 @@ class CourseCreateController extends Controller
         });
 
         $course_draft = auth()->user()->courseDraft->last();
-        if($course === null && $course_draft) {
+        if ($course === null && $course_draft) {
             $course = [
                 'draft_id'    => $course_draft->id,
                 'title'       => $course_draft['title'],
@@ -424,17 +409,13 @@ class CourseCreateController extends Controller
 
 
             $draft_id = ParameterService::get($request->toArray(),'draft_id', null);
-            if ($draft_id) {
-                CourseDraft::where('id',$draft_id)->delete();
-            }
+            if ($draft_id) CourseDraft::where('id',$draft_id)->delete();
 
             DB::commit();
-
         } catch (\Exception $e) {
             DB::rollback();
             throw $e;
         }
-
         $request->session()->forget('course');
         $request->session()->forget('save_teacher');
         $request->session()->forget('signed_contract');
